Tests/LibWeb/Text/expected/css/getComputedStyle-blockification-on-parent-display-change.txt

Initial child display: inline
Color after flex: rgb(0, 0, 0)
Display after flex: block
Display after block: inline
